requirements.txt
vosk
sounddevice
openai
faster-whisper
numpy
pydub
pyaudio
coqui-tts
SQLAlchemy
psycopg2-binary
alembic
python-dotenv